现在全部搞清楚了。Session ID 的结构是 gate_node_id << 32 | seq，日志中所有 session_list 值都是 4294967xxx（即 0x1_0000_xxxx），说明它们共享同一个 gate。
关键发现：同一个 gate 的广播消息里，所有 session ID 的高 32 位完全相同。每个 session ID 用 varint 编码需要 ~5 字节，但实际上只有低 32 位不同。

优化方案：session_list 只传低 32 位（repeated uint32），gate_node_id 已经隐含在"这条 RPC 发给了哪个 gate"中。Gate 收到后重组 session ID。

这样每个 session 从 ~5 字节变成 ~2-3 字节，session_list 大小减半。

+++ cpp/libs/engine/core/network/player_message_utils.cpp
@@ -131,7 +131,7 @@ void InternalBroadcast(uint32_t messageId, const google::protobuf::Message &mess
 
 		for (auto &&sessionId : sessionIdList)
 		{
-			request.mutable_session_list()->Add(sessionId);
+			request.mutable_session_list()->Add(static_cast<uint32_t>(sessionId));
 		}
 
 		gateNodeSession->SendRequest(GateBroadcastToPlayersMessageId, request);
